Move convertQuality to shared-schema, remove duplication
The function was duplicated in aurora-sync and web packages with no
shared coverage. Move the single source of truth to
@boardsesh/shared-schema/utils and have both packages import from there.
