+++ packages/aurora-sync/package.json
@@ -42,6 +42,7 @@
   "dependencies": {
     "@boardsesh/crypto": "*",
     "@boardsesh/db": "*",
+    "@boardsesh/shared-schema": "*",
     "@neondatabase/serverless": "1.0.2",
     "commander": "^14.0.3",
     "dotenv": "^17.3.1",
