+++ packages/aurora-sync/src/sync/index.ts
@@ -1,3 +1,3 @@
 export { syncUserData, getLastSyncTimes, getLastSharedSyncTimes } from './user-sync';
 export type { SyncUserDataResult } from './user-sync';
-export { convertQuality } from './convert-quality';
+export { convertQuality } from '@boardsesh/shared-schema';
