+++ packages/shared-schema/src/utils.ts
@@ -1,3 +1,20 @@
 export function isNoMatchClimb(description: string | null | undefined): boolean {
   return /^no match/i.test(description || '');
 }
+
+/**
+ * Convert an Aurora quality rating (1-3) to a Boardsesh quality rating (1-5).
+ *
+ * Aurora's Kilter/Tension logbook stores user star ratings on a 1-3 scale
+ * (0 means "unrated" and maps to null). Boardsesh stores them on a 1-5 scale.
+ * We map endpoints exactly (1->1, 3->5) with 2->3 in the middle via linear
+ * interpolation, and clamp defensively in case Aurora ever returns values
+ * outside 1-3.
+ */
+export function convertQuality(auroraQuality: number | null | undefined): number | null {
+  if (auroraQuality == null) return null;
+  const q = Number(auroraQuality);
+  if (!Number.isFinite(q) || q <= 0) return null;
+  const clamped = Math.min(3, Math.max(1, q));
+  return Math.round(((clamped - 1) / 2) * 4) + 1;
+}
